backend/requirements.txt
flask
flask-cors
requests
python-dotenv
google-api-python-client